Move profile model to accounts app and delete users app.
