+++ govapp/apps/accounts/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.contrib.auth.models import User
 
-from govapp.apps.users.models import Profile
+from govapp.apps.accounts.models import Profile
 
 
 # Define an inline admin descriptor for Employee model
